requirements.txt
PySide6
PyOpenGL
PyOpenGL_accelerate
ezdxf
numpy